requirements.txt
Django==3.1
psycopg2-binary>=2.8
djangorestframework
django-allauth==0.44.0
django-guardian
django-widget-tweaks
django-phonenumber-field[phonenumbers]
Babel
Pillow
gunicorn